+++ CHANGELOG.md
@@ -55,6 +55,7 @@ doivent être mises à jour #1667
 - Suppression du script `optimize.php`, remplacé par la commande app:database:optimize #2293
 - Suppression du script `cron-jour.php`, remplacé par la commande app:notification:daily-digest #2293
 - Suppression du script `reindex-documents.php`, remplacé par la commande `app:module:reindex` #2297
+- Suppression du script `fix-module.php`, remplacé par la commande `app:module:change-type` #2295
 - Suppression du script `delete-association-authentification.php`, remplacé par la commande `app:connector:dissociate` #2298
 
 ## Suppressions d'éléments dépréciés